/feature_point_tags add tag delete, swap point id type to long, insert details entity after point creation

+++ app/src/main/java/com/example/gh_coursework/data/database/entity/PointDetailsEntity.kt
@@ -14,7 +14,7 @@ import androidx.room.PrimaryKey
 data class PointDetailsEntity(
     @PrimaryKey
     @ColumnInfo(name = "pointId")
-    val pointId: Int,
+    val pointId: Long,
     @ColumnInfo(name = "caption")
     val caption: String,
     @ColumnInfo(name = "description")
