+++ app/src/main/java/com/example/gh_coursework/domain/entity/PointDetailsDomain.kt
@@ -1,7 +1,7 @@
 package com.example.gh_coursework.domain.entity
 
 data class PointDetailsDomain(
-    val pointId: Int,
+    val pointId: Long,
     val tagList: List<PointTagDomain>,
     val caption: String,
     val description: String
