+++ app/src/main/java/com/example/gh_coursework/ui/point_details/model/PointDetailsModel.kt
@@ -1,7 +1,7 @@
 package com.example.gh_coursework.ui.point_details.model
 
 data class PointDetailsModel(
-    val pointId: Int,
+    val pointId: Long,
     val tagList: List<PointTagModel>,
     val caption: String,
     val description: String
