+++ app/src/main/java/com/example/gh_coursework/ui/point_details/model/PointsTagsModel.kt
@@ -1,6 +1,6 @@
 package com.example.gh_coursework.ui.point_details.model
 
 data class PointsTagsModel(
-    val pointId: Int,
+    val pointId: Long,
     val tagId: Int
 )
